backend/requirements.txt
Django==4.1
gunicorn==19.6.0
djangorestframework==3.14.0
drf-spectacular==0.24.1
black==22.12.0

-e .

-e ./science
 